+++ src/omnid.ts
@@ -1,11 +1,12 @@
 import { fetcher } from './utils';
 import { ComputeConfig, ErrorType } from './types';
-import * as adaptors from './adaptors';
+import * as adaptorList from './adaptors';
 import { isAddress } from 'ethers/lib/utils';
 
 class Identity {
   apikey: string;
   node: string;
+  adaptors = adaptorList;
 
   constructor(apikey: string, node: string) {
     this.apikey = apikey;
@@ -28,36 +29,38 @@ class Identity {
   ): Promise<any | ErrorType> => {
     if (isAddress(address) === true) {
       const promiseArray = [
-        adaptors.getAaveData(address, computeConfig),
-        adaptors.getAge(address),
-        adaptors.getArcxData(address),
-        adaptors.getAsyncartData(address),
-        adaptors.getBoardroomData(address),
-        adaptors.checkBrightId(address),
-        adaptors.getCeloData(address),
-        adaptors.getContextData(address),
-        // adaptors.getCoordinapeData(address),
-        adaptors.getCryptoscamdbData(address),
-        adaptors.getCyberconnectData(address),
-        adaptors.getDeepDaoData(address, computeConfig),
-        adaptors.addressToEns(address),
-        adaptors.getFoundationData(address),
-        adaptors.getGitcoinData(address, computeConfig),
-        adaptors.checkIdena(address),
-        adaptors.getKnownOriginData(address),
-        adaptors.getMetagameData(address),
-        adaptors.getMirrorData(address),
-        adaptors.getPoapData(address),
-        adaptors.getPolygonData(address),
-        adaptors.getProjectGalaxyData(address),
-        adaptors.checkPoH(address),
-        adaptors.getRabbitholeData(address),
-        adaptors.getRaribleData(address),
-        adaptors.getRss3Data(address),
-        adaptors.getShowtimeData(address, computeConfig),
-        adaptors.getSuperrareData(address),
-        adaptors.resolveUnstoppableDomains(address),
-        adaptors.getZoraData(address),
+        adaptorList.getAaveData(address, computeConfig),
+        adaptorList.getAge(address),
+        adaptorList.getArcxData(address),
+        adaptorList.getAsyncartData(address),
+        adaptorList.getBoardroomData(address),
+        adaptorList.checkBrightId(address),
+        adaptorList.getCeloData(address),
+        adaptorList.getCoinviseData(address, computeConfig),
+        adaptorList.getContextData(address),
+        // adaptorList.getCoordinapeData(address),
+        adaptorList.getCryptoscamdbData(address),
+        adaptorList.getCyberconnectData(address),
+        adaptorList.getDeepDaoData(address, computeConfig),
+        adaptorList.addressToEns(address),
+        adaptorList.getFoundationData(address),
+        adaptorList.getGitcoinData(address, computeConfig),
+        adaptorList.checkIdena(address),
+        adaptorList.getKnownOriginData(address),
+        adaptorList.getMetagameData(address),
+        adaptorList.getMirrorData(address),
+        adaptorList.getPoapData(address),
+        adaptorList.getPolygonData(address),
+        adaptorList.getProjectGalaxyData(address),
+        adaptorList.checkPoH(address),
+        adaptorList.getRabbitholeData(address),
+        adaptorList.getRaribleData(address),
+        adaptorList.getRss3Data(address),
+        adaptorList.getShowtimeData(address, computeConfig),
+        adaptorList.getSuperrareData(address),
+        adaptorList.getSybilData(address, computeConfig),
+        adaptorList.resolveUnstoppableDomains(address),
+        adaptorList.getZoraData(address),
       ];
       if (Boolean(computeConfig?.DEBUG) === true) console.time('computeTime');
       const resp: Array<PromiseSettledResult<any>> = await Promise.allSettled(
@@ -71,5 +74,5 @@ class Identity {
     }
   };
 }
-export { adaptors };
+
 export default Identity;